chore: move error handling to a middlewares folder
