+++ src/api/server.ts
@@ -1,8 +1,7 @@
-import express, { NextFunction, Request, Response } from 'express'
+import express from 'express'
 import bodyParser from 'body-parser'
 import { devices } from './devices/devices'
-import { DomainError, ErrorCategory } from '../shared/domain/errors/DomainError'
-import { config } from '../config'
+import { errorHandler } from './middlewares/errorHandler'
 
 const server = express()
 
@@ -19,29 +18,6 @@ server.get('/health', (req, res, next) => {
 })
 
 server.use('/devices', devices)
-
-const errorCategoryToHttpCode: Record<ErrorCategory, number> = {
-  [ErrorCategory.BAD_REQUEST]: 400,
-  [ErrorCategory.UNKNOWN]: 500,
-}
-
-const errorHandler = (err: Error, req: Request, res: Response, next: NextFunction) => {
-  if (err instanceof DomainError) {
-    res.status(errorCategoryToHttpCode[err.category])
-    res.send({
-      category: err.category,
-      message: err.message,
-      stack: config.node.env === 'development' ? err.stack : undefined,
-    })
-  }
-
-  res.status(500)
-  res.send({
-    category: ErrorCategory.UNKNOWN,
-    message: config.node.env === 'development' ? err.message : 'Something went wrong',
-    stack: config.node.env === 'development' ? err.stack : undefined,
-  })
-}
 
 server.use(errorHandler)
